requirements.txt
--index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
torchvision==0.16.0
pillow==10.2.0
